# Kannada-Speech-Aid/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kannada Speech Therapy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type ExerciseCategory =
  (typeof ExerciseCategory)[keyof typeof ExerciseCategory];

export const ExerciseCategory = {
  vowels: "vowels",
  consonants: "consonants",
  words: "words",
  sentences: "sentences",
} as const;

export type ExerciseDifficulty =
  (typeof ExerciseDifficulty)[keyof typeof ExerciseDifficulty];

export const ExerciseDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface Exercise {
  id: number;
  kannadaText: string;
  transliteration: string;
  englishMeaning: string;
  category: ExerciseCategory;
  difficulty: ExerciseDifficulty;
  audioPrompt?: string | null;
  orderIndex: number;
}

export interface Patient {
  id: number;
  name: string;
  age: number;
  condition: string;
  therapistNotes?: string | null;
  createdAt: string;
  totalSessions: number;
  averageScore?: number | null;
}

export interface CreatePatientBody {
  name: string;
  age: number;
  condition: string;
  therapistNotes?: string;
}

export type TherapySessionStatus =
  (typeof TherapySessionStatus)[keyof typeof TherapySessionStatus];

export const TherapySessionStatus = {
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface TherapySession {
  id: number;
  patientId: number;
  startedAt: string;
  completedAt?: string | null;
  totalAttempts: number;
  averageScore?: number | null;
  status: TherapySessionStatus;
}

export type TherapySessionWithAttemptsStatus =
  (typeof TherapySessionWithAttemptsStatus)[keyof typeof TherapySessionWithAttemptsStatus];

export const TherapySessionWithAttemptsStatus = {
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface SpeechAttempt {
  id: number;
  sessionId: number;
  exerciseId: number;
  spokenText: string;
  accuracyScore: number;
  feedbackText: string;
  attemptedAt: string;
}

export interface TherapySessionWithAttempts {
  id: number;
  patientId: number;
  startedAt: string;
  completedAt?: string | null;
  totalAttempts: number;
  averageScore?: number | null;
  status: TherapySessionWithAttemptsStatus;
  attempts: SpeechAttempt[];
}

export interface CreateSessionBody {
  patientId: number;
}

export type UpdateSessionBodyStatus =
  (typeof UpdateSessionBodyStatus)[keyof typeof UpdateSessionBodyStatus];

export const UpdateSessionBodyStatus = {
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface UpdateSessionBody {
  status?: UpdateSessionBodyStatus;
}

export interface RecordAttemptBody {
  exerciseId: number;
  spokenText: string;
  accuracyScore: number;
  feedbackText: string;
}

export interface TranscribeSpeechBody {
  /** Base64-encoded audio data */
  audio: string;
  /** The Kannada text the patient was supposed to say */
  expectedText: string;
}

export interface TranscribeSpeechResponse {
  transcribedText: string;
  accuracyScore: number;
  feedbackText: string;
  suggestions: string[];
}

export interface WeeklyScore {
  week: string;
  averageScore: number;
  sessionsCount: number;
}

export interface CategoryScore {
  category: string;
  averageScore: number;
  totalAttempts: number;
}

export interface PatientProgress {
  patientId: number;
  totalSessions: number;
  totalAttempts: number;
  averageScore?: number | null;
  weeklyScores: WeeklyScore[];
  categoryScores: CategoryScore[];
}

export type PatientProgressSummaryTrend =
  (typeof PatientProgressSummaryTrend)[keyof typeof PatientProgressSummaryTrend];

export const PatientProgressSummaryTrend = {
  improving: "improving",
  stable: "stable",
  declining: "declining",
  no_data: "no_data",
} as const;

export interface PatientProgressSummary {
  patientId: number;
  patientName: string;
  totalSessions: number;
  lastSessionDate?: string | null;
  averageScore?: number | null;
  trend: PatientProgressSummaryTrend;
}

export type ListExercisesParams = {
  category?: ListExercisesCategory;
  difficulty?: ListExercisesDifficulty;
};

export type ListExercisesCategory =
  (typeof ListExercisesCategory)[keyof typeof ListExercisesCategory];

export const ListExercisesCategory = {
  vowels: "vowels",
  consonants: "consonants",
  words: "words",
  sentences: "sentences",
} as const;

export type ListExercisesDifficulty =
  (typeof ListExercisesDifficulty)[keyof typeof ListExercisesDifficulty];

export const ListExercisesDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;