# Kannada-Speech-Aid/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Kannada Speech Therapy API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: exercises
    description: Speech therapy exercises
  - name: sessions
    description: Therapy sessions and recordings
  - name: patients
    description: Patient management
  - name: progress
    description: Progress tracking
  - name: speech
    description: Speech recognition and AI feedback
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /exercises:
    get:
      operationId: listExercises
      tags: [exercises]
      summary: List all exercises
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [vowels, consonants, words, sentences]
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [beginner, intermediate, advanced]
      responses:
        "200":
          description: List of exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"

  /exercises/{id}:
    get:
      operationId: getExercise
      tags: [exercises]
      summary: Get a single exercise
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exercise details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List all patients
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Register a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientBody"
      responses:
        "201":
          description: Created patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"

  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get patient details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /patients/{id}/sessions:
    get:
      operationId: getPatientSessions
      tags: [patients]
      summary: Get therapy sessions for a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TherapySession"

  /patients/{id}/progress:
    get:
      operationId: getPatientProgress
      tags: [patients]
      summary: Get progress data for a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Progress data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientProgress"

  /sessions:
    post:
      operationId: createSession
      tags: [sessions]
      summary: Start a new therapy session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionBody"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TherapySession"

  /sessions/{id}:
    get:
      operationId: getSession
      tags: [sessions]
      summary: Get session details with attempts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TherapySessionWithAttempts"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateSession
      tags: [sessions]
      summary: Update session (e.g. mark complete)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSessionBody"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TherapySession"

  /sessions/{id}/attempts:
    post:
      operationId: recordAttempt
      tags: [sessions]
      summary: Record a speech attempt for an exercise
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordAttemptBody"
      responses:
        "201":
          description: Recorded attempt with AI feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeechAttempt"

  /speech/transcribe:
    post:
      operationId: transcribeSpeech
      tags: [speech]
      summary: Transcribe recorded speech audio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscribeSpeechBody"
      responses:
        "200":
          description: Transcription result with AI feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscribeSpeechResponse"

  /progress/summary:
    get:
      operationId: getProgressSummary
      tags: [progress]
      summary: Get overall progress summary for all patients
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PatientProgressSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Exercise:
      type: object
      properties:
        id:
          type: integer
        kannadaText:
          type: string
        transliteration:
          type: string
        englishMeaning:
          type: string
        category:
          type: string
          enum: [vowels, consonants, words, sentences]
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        audioPrompt:
          type: string
          nullable: true
        orderIndex:
          type: integer
      required:
        - id
        - kannadaText
        - transliteration
        - englishMeaning
        - category
        - difficulty
        - orderIndex

    Patient:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        age:
          type: integer
        condition:
          type: string
        therapistNotes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        totalSessions:
          type: integer
        averageScore:
          type: number
          nullable: true
      required:
        - id
        - name
        - age
        - condition
        - createdAt
        - totalSessions

    CreatePatientBody:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
        condition:
          type: string
        therapistNotes:
          type: string
      required:
        - name
        - age
        - condition

    TherapySession:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        totalAttempts:
          type: integer
        averageScore:
          type: number
          nullable: true
        status:
          type: string
          enum: [in_progress, completed]
      required:
        - id
        - patientId
        - startedAt
        - totalAttempts
        - status

    TherapySessionWithAttempts:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        totalAttempts:
          type: integer
        averageScore:
          type: number
          nullable: true
        status:
          type: string
          enum: [in_progress, completed]
        attempts:
          type: array
          items:
            $ref: "#/components/schemas/SpeechAttempt"
      required:
        - id
        - patientId
        - startedAt
        - totalAttempts
        - status
        - attempts

    CreateSessionBody:
      type: object
      properties:
        patientId:
          type: integer
      required:
        - patientId

    UpdateSessionBody:
      type: object
      properties:
        status:
          type: string
          enum: [in_progress, completed]

    SpeechAttempt:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        exerciseId:
          type: integer
        spokenText:
          type: string
        accuracyScore:
          type: number
        feedbackText:
          type: string
        attemptedAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - exerciseId
        - spokenText
        - accuracyScore
        - feedbackText
        - attemptedAt

    RecordAttemptBody:
      type: object
      properties:
        exerciseId:
          type: integer
        spokenText:
          type: string
        accuracyScore:
          type: number
        feedbackText:
          type: string
      required:
        - exerciseId
        - spokenText
        - accuracyScore
        - feedbackText

    TranscribeSpeechBody:
      type: object
      properties:
        audio:
          type: string
          description: Base64-encoded audio data
        expectedText:
          type: string
          description: The Kannada text the patient was supposed to say
      required:
        - audio
        - expectedText

    TranscribeSpeechResponse:
      type: object
      properties:
        transcribedText:
          type: string
        accuracyScore:
          type: number
        feedbackText:
          type: string
        suggestions:
          type: array
          items:
            type: string
      required:
        - transcribedText
        - accuracyScore
        - feedbackText
        - suggestions

    PatientProgress:
      type: object
      properties:
        patientId:
          type: integer
        totalSessions:
          type: integer
        totalAttempts:
          type: integer
        averageScore:
          type: number
          nullable: true
        weeklyScores:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyScore"
        categoryScores:
          type: array
          items:
            $ref: "#/components/schemas/CategoryScore"
      required:
        - patientId
        - totalSessions
        - totalAttempts
        - weeklyScores
        - categoryScores

    WeeklyScore:
      type: object
      properties:
        week:
          type: string
        averageScore:
          type: number
        sessionsCount:
          type: integer
      required:
        - week
        - averageScore
        - sessionsCount

    CategoryScore:
      type: object
      properties:
        category:
          type: string
        averageScore:
          type: number
        totalAttempts:
          type: integer
      required:
        - category
        - averageScore
        - totalAttempts

    PatientProgressSummary:
      type: object
      properties:
        patientId:
          type: integer
        patientName:
          type: string
        totalSessions:
          type: integer
        lastSessionDate:
          type: string
          format: date-time
          nullable: true
        averageScore:
          type: number
          nullable: true
        trend:
          type: string
          enum: [improving, stable, declining, no_data]
      required:
        - patientId
        - patientName
        - totalSessions
        - trend