# Kannada-Speech-Aid/lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kannada Speech Therapy API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all exercises
 */
export const ListExercisesQueryParams = zod.object({
  category: zod.enum(["vowels", "consonants", "words", "sentences"]).optional(),
  difficulty: zod.enum(["beginner", "intermediate", "advanced"]).optional(),
});

export const ListExercisesResponseItem = zod.object({
  id: zod.number(),
  kannadaText: zod.string(),
  transliteration: zod.string(),
  englishMeaning: zod.string(),
  category: zod.enum(["vowels", "consonants", "words", "sentences"]),
  difficulty: zod.enum(["beginner", "intermediate", "advanced"]),
  audioPrompt: zod.string().nullish(),
  orderIndex: zod.number(),
});
export const ListExercisesResponse = zod.array(ListExercisesResponseItem);

/**
 * @summary Get a single exercise
 */
export const GetExerciseParams = zod.object({
  id: zod.coerce.number(),
});

export const GetExerciseResponse = zod.object({
  id: zod.number(),
  kannadaText: zod.string(),
  transliteration: zod.string(),
  englishMeaning: zod.string(),
  category: zod.enum(["vowels", "consonants", "words", "sentences"]),
  difficulty: zod.enum(["beginner", "intermediate", "advanced"]),
  audioPrompt: zod.string().nullish(),
  orderIndex: zod.number(),
});

/**
 * @summary List all patients
 */
export const ListPatientsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  age: zod.number(),
  condition: zod.string(),
  therapistNotes: zod.string().nullish(),
  createdAt: zod.date(),
  totalSessions: zod.number(),
  averageScore: zod.number().nullish(),
});
export const ListPatientsResponse = zod.array(ListPatientsResponseItem);

/**
 * @summary Register a new patient
 */
export const CreatePatientBody = zod.object({
  name: zod.string(),
  age: zod.number(),
  condition: zod.string(),
  therapistNotes: zod.string().optional(),
});

/**
 * @summary Get patient details
 */
export const GetPatientParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPatientResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  age: zod.number(),
  condition: zod.string(),
  therapistNotes: zod.string().nullish(),
  createdAt: zod.date(),
  totalSessions: zod.number(),
  averageScore: zod.number().nullish(),
});

/**
 * @summary Get therapy sessions for a patient
 */
export const GetPatientSessionsParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPatientSessionsResponseItem = zod.object({
  id: zod.number(),
  patientId: zod.number(),
  startedAt: zod.date(),
  completedAt: zod.date().nullish(),
  totalAttempts: zod.number(),
  averageScore: zod.number().nullish(),
  status: zod.enum(["in_progress", "completed"]),
});
export const GetPatientSessionsResponse = zod.array(
  GetPatientSessionsResponseItem,
);

/**
 * @summary Get progress data for a patient
 */
export const GetPatientProgressParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPatientProgressResponse = zod.object({
  patientId: zod.number(),
  totalSessions: zod.number(),
  totalAttempts: zod.number(),
  averageScore: zod.number().nullish(),
  weeklyScores: zod.array(
    zod.object({
      week: zod.string(),
      averageScore: zod.number(),
      sessionsCount: zod.number(),
    }),
  ),
  categoryScores: zod.array(
    zod.object({
      category: zod.string(),
      averageScore: zod.number(),
      totalAttempts: zod.number(),
    }),
  ),
});

/**
 * @summary Start a new therapy session
 */
export const CreateSessionBody = zod.object({
  patientId: zod.number(),
});

/**
 * @summary Get session details with attempts
 */
export const GetSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetSessionResponse = zod.object({
  id: zod.number(),
  patientId: zod.number(),
  startedAt: zod.date(),
  completedAt: zod.date().nullish(),
  totalAttempts: zod.number(),
  averageScore: zod.number().nullish(),
  status: zod.enum(["in_progress", "completed"]),
  attempts: zod.array(
    zod.object({
      id: zod.number(),
      sessionId: zod.number(),
      exerciseId: zod.number(),
      spokenText: zod.string(),
      accuracyScore: zod.number(),
      feedbackText: zod.string(),
      attemptedAt: zod.date(),
    }),
  ),
});

/**
 * @summary Update session (e.g. mark complete)
 */
export const UpdateSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateSessionBody = zod.object({
  status: zod.enum(["in_progress", "completed"]).optional(),
});

export const UpdateSessionResponse = zod.object({
  id: zod.number(),
  patientId: zod.number(),
  startedAt: zod.date(),
  completedAt: zod.date().nullish(),
  totalAttempts: zod.number(),
  averageScore: zod.number().nullish(),
  status: zod.enum(["in_progress", "completed"]),
});

/**
 * @summary Record a speech attempt for an exercise
 */
export const RecordAttemptParams = zod.object({
  id: zod.coerce.number(),
});

export const RecordAttemptBody = zod.object({
  exerciseId: zod.number(),
  spokenText: zod.string(),
  accuracyScore: zod.number(),
  feedbackText: zod.string(),
});

/**
 * @summary Transcribe recorded speech audio
 */
export const TranscribeSpeechBody = zod.object({
  audio: zod.string().describe("Base64-encoded audio data"),
  expectedText: zod
    .string()
    .describe("The Kannada text the patient was supposed to say"),
});

export const TranscribeSpeechResponse = zod.object({
  transcribedText: zod.string(),
  accuracyScore: zod.number(),
  feedbackText: zod.string(),
  suggestions: zod.array(zod.string()),
});

/**
 * @summary Get overall progress summary for all patients
 */
export const GetProgressSummaryResponseItem = zod.object({
  patientId: zod.number(),
  patientName: zod.string(),
  totalSessions: zod.number(),
  lastSessionDate: zod.date().nullish(),
  averageScore: zod.number().nullish(),
  trend: zod.enum(["improving", "stable", "declining", "no_data"]),
});
export const GetProgressSummaryResponse = zod.array(
  GetProgressSummaryResponseItem,
);