# Kannada-Speech-Aid/lib/api-zod/src/generated/types/createPatientBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kannada Speech Therapy API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePatientBody {
  name: string;
  age: number;
  condition: string;
  therapistNotes?: string;
}