# Kannada-Speech-Aid/lib/api-zod/src/generated/types/exercise.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kannada Speech Therapy API
 * OpenAPI spec version: 0.1.0
 */
import type { ExerciseCategory } from "./exerciseCategory";
import type { ExerciseDifficulty } from "./exerciseDifficulty";

export interface Exercise {
  id: number;
  kannadaText: string;
  transliteration: string;
  englishMeaning: string;
  category: ExerciseCategory;
  difficulty: ExerciseDifficulty;
  audioPrompt?: string | null;
  orderIndex: number;
}