# Kannada-Speech-Aid/lib/api-zod/src/generated/types/patientProgress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kannada Speech Therapy API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryScore } from "./categoryScore";
import type { WeeklyScore } from "./weeklyScore";

export interface PatientProgress {
  patientId: number;
  totalSessions: number;
  totalAttempts: number;
  averageScore?: number | null;
  weeklyScores: WeeklyScore[];
  categoryScores: CategoryScore[];
}